finished jwt

+++ Saitynai/CharacterEndpoints.cs
@@ -2,6 +2,14 @@
 using Saitynai.Data;
 using Microsoft.EntityFrameworkCore;
 using SharpGrip.FluentValidation.AutoValidation.Endpoints.Extensions;
+using Microsoft.AspNetCore.Http;
+using Saitynai.Auth.Model;
+using System.Net.Http;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.IdentityModel.Tokens;
+using System.Security.Claims;
+using System.IdentityModel.Tokens.Jwt;
+using Microsoft.AspNetCore.Http;
 
 namespace Saitynai
 {
@@ -24,12 +32,18 @@ namespace Saitynai
                 return character == null ? Results.NotFound() : Results.Ok(character.ToDto());
             });
 
-            charactersGroup.MapPost("/characters/", async (int scriptId, CreateCharacterDto dto, ForumDbContext dbContext) =>
+            charactersGroup.MapPost("/characters/", [Authorize(Roles = ForumRoles.ForumUser)] async (int scriptId, CreateCharacterDto dto, ForumDbContext dbContext, HttpContext httpContext) =>
             {
                 var script = await dbContext.Scripts.FindAsync(scriptId);
                 if (script == null)
                 {
                     return Results.NotFound();
+                }
+
+                if (!httpContext.User.IsInRole(ForumRoles.Admin) && httpContext.User.FindFirstValue(JwtRegisteredClaimNames.Sub) != script.UserId)
+                {
+                    //NotFound()
+                    return Results.Forbid();
                 }
 
                 var character = new Character
@@ -38,7 +52,7 @@ namespace Saitynai
                     Body = dto.Body,
                     CreatedAt = DateTimeOffset.UtcNow,
                     Script = script,
-                    UserId = ""
+                    UserId = httpContext.User.FindFirstValue(JwtRegisteredClaimNames.Sub)
                 };
                 dbContext.Characters.Add(character);
                 await dbContext.SaveChangesAsync();
@@ -46,12 +60,18 @@ namespace Saitynai
                 return TypedResults.Created($"/api/scripts/{scriptId}/characters/{character.Id}", character.ToDto());
             });
 
-            charactersGroup.MapPut("/characters/{characterId:int}", async (int scriptId, int characterId, UpdateCharacterDto dto, ForumDbContext dbContext) =>
+            charactersGroup.MapPut("/characters/{characterId:int}", [Authorize] async (int scriptId, int characterId, UpdateCharacterDto dto, ForumDbContext dbContext, HttpContext httpContext) =>
             {
                 var character = await dbContext.Characters.FirstOrDefaultAsync(c => c.Id == characterId && c.Script.Id == scriptId);
                 if (character == null)
                 {
                     return Results.NotFound();
+                }
+
+                if (!httpContext.User.IsInRole(ForumRoles.Admin) && httpContext.User.FindFirstValue(JwtRegisteredClaimNames.Sub) != character.UserId)
+                {
+                    //NotFound()
+                    return Results.Forbid();
                 }
 
                 character.Body = dto.Body;
@@ -62,12 +82,17 @@ namespace Saitynai
                 return TypedResults.Ok(character.ToDto());
             });
 
-            charactersGroup.MapDelete("/characters/{characterId:int}", async (int scriptId, int characterId, ForumDbContext dbContext) =>
+            charactersGroup.MapDelete("/characters/{characterId:int}", [Authorize] async (int scriptId, int characterId, ForumDbContext dbContext, HttpContext httpContext) =>
             {
                 var character = await dbContext.Characters.FirstOrDefaultAsync(c => c.Id == characterId && c.Script.Id == scriptId);
                 if (character == null)
                 {
                     return Results.NotFound();
+                }
+                if (!httpContext.User.IsInRole(ForumRoles.Admin) && httpContext.User.FindFirstValue(JwtRegisteredClaimNames.Sub) != character.UserId)
+                {
+                    //NotFound()
+                    return Results.Forbid();
                 }
 
                 dbContext.Characters.Remove(character);
